requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
langgraph
mcp
httpx
python-dotenv
langchain-openai
langchain-community
langchain-core